# src/main/resources/META-INF/openapi.yaml
openapi: 3.0.3
info:
  title: Authorization Microservice
  version: 1.10.0
  description: Authorization Microservice for Multi-channel Integration Layer of SW Client Project
  contact:
    name: Antonio Tarricone
    email: antonio.tarricone@pagopa.it
servers:
  - description: Development/Test
    url: https://mil-d-apim.azure-api.net/mil-auth
    x-internal: true
  - description: User Acceptance Test
    url: https://mil-u-apim.azure-api.net/mil-auth
    x-internal: true
...
paths:
  /token:
    post:
      operationId: getAccessToken
      description: Returns access token or refresh it
      tags: [ token ]
      security: [ ]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Version'
        - $ref: '#/components/parameters/AcquirerId'
        - $ref: '#/components/parameters/Channel'
        - $ref: '#/components/parameters/MerchantId'
        - $ref: '#/components/parameters/TerminalId'
        - $ref: '#/components/parameters/FiscalCode'
      requestBody:
        $ref: '#/components/requestBodies/GetAccessToken'
      responses:
        "201":
          #description: Created
          $ref: '#/components/responses/AccessToken'
        "400":
          #description: Bad request
          $ref: '#/components/responses/Error'
        "401":
          #description: Wrong credentials
          $ref: '#/components/responses/Error'
        "403":
          #description: Forbidden
          $ref: '#/components/responses/Error'
        "406":
          #description: Not acceptable. Did you require application/json?
          $ref: '#/components/responses/Error'
        "415":
          #description: Unsupported media type. Did you provide application/x-www-form-urlencoded?
          $ref: '#/components/responses/Error'
        "429":
          #description: Too many request
          $ref: '#/components/responses/Error'
        "500":
          #description: Server error
          $ref: '#/components/responses/Error'
        default:
          #description: Unexpected error
          $ref: '#/components/responses/Error'
...
  /admin/clients:
    post:
      operationId: createClient
      description: Creates a new client
      tags: [ admin ]
      security:
        - oAuth2: [ admin ]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Version'
      requestBody:
        $ref: '#/components/requestBodies/CreateClient'
      responses:
        "201":
          #description: Created
          $ref: '#/components/responses/CreateClient'
        "401":
          #description: Wrong credentials
          $ref: '#/components/responses/Error'
        "403":
          #description: Forbidden
          $ref: '#/components/responses/Error'
        "406":
          #description: Not acceptable. Did you require application/json?
          $ref: '#/components/responses/Error'
        "415":
          #description: Unsupported media type. Did you provide application/json?
          $ref: '#/components/responses/Error'
        "429":
          #description: Too many request
          $ref: '#/components/responses/Error'
        "500":
          #description: Server error
          $ref: '#/components/responses/Error'
        default:
          #description: Unexpected error
          $ref: '#/components/responses/Error'

    get:
      operationId: getClients
      description: Returns all clients
      tags: [ admin ]
      security:
        - oAuth2: [ admin ]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Version'
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        "200":
          #description: Ok
          $ref: '#/components/responses/PageOfClients'
        "400":
          #description: Bad request
          $ref: '#/components/responses/Error'
        "401":
          #description: Access token is missing or invalid
          $ref: '#/components/responses/Error'
        "403":
          #description: Forbidden
          $ref: '#/components/responses/Error'
        "406":
          #description: Not acceptable. Did you require application/json?
          $ref: '#/components/responses/Error'
        "429":
          #description: Too many request
          $ref: '#/components/responses/Error'
        "500":
          #description: Server error
          $ref: '#/components/responses/Error'
        default:
          description: Unexpected error

  
  #/admin/roles:
components:
...
  schemas:
...
    AccessControlAllowOrigin:
      description: Indicates whether the response can be shared with requesting code from the given origin
      type: string
      pattern: "^[ -~]{1,2048}$"
      minLength: 1
      maxLength: 2048
    AcquirerId:
      description: Acquirer ID assigned by PagoPA
      type: string
      pattern: "^\\d{1,11}$"
      minLength: 1
      maxLength: 11
      example: "4585625"
...
    Channel:
      description: Channel originating the request
      enum:
        - ATM
        - POS
        - TOTEM
        - CASH_REGISTER
        - CSA
      type: string
      example: "ATM"
...
    ClientLocation:
      description: URL of the client
      type: string
      format: uri
      pattern: "^[ -~]{1,2048}$"
      minLength: 1
      maxLength: 2048
      example: "https://mil-d-apim.azure-api.net/mil-auth/admin/clients/d0d654e6-97da-4848-b568-99fedccb642b"
...
    FiscalCode:
      description: User fiscal (tax) code
      type: string
      pattern: "^(([A-Z]{6}\\d{2}[A-Z]\\d{2}[A-Z]\\d{3}[A-Z])|(\\d{11}))$"
      minLength: 11
      maxLength: 16
      example: "RSSMRA85T10A562S"
...
    MerchantId:
      description: Merchant ID. Mandatory when Channel equals POS.
      type: string
      pattern: "^[0-9a-zA-Z]{1,15}$"
      minLength: 1
      maxLength: 15
      example: "28405fHfk73x88D"   
...
    PageNumber:
      description: Number of the page
      type: integer
      format: int32 # TotalPages and TotalElements are int64 but PageNumber is int32 due to Panache APIs limits.
      minimum: 1
      maximum: 2147483647
      example: 1
    PageSize:
      description: Size of the page
      type: integer
      format: int32
      minimum: 1
      maximum: 128 # See #/components/schemas/Clients.maxItems
      example: 20
    PasswordGrantType:
      description: Grant type to get an access token with resource owner password
      enum:
        - password
      type: string
      example: "password"
...
    RateLimitLimit:
      description: The number of allowed requests in the current period
      type: integer
      format: int32
      minimum: 1
      maximum: 240
    RateLimitReset:
      description: The number of seconds left in the current period
      type: integer
      format: int32
      minimum: 1
      maximum: 60
    RetryAfter:
      description: The number of seconds to wait before allowing a follow-up request
      type: integer
      format: int32
      minimum: 1
      maximum: 240
...
    TerminalId:
      description: ID of the terminal originating the transaction. It must be unique per acquirer and channel.
      type: string
      pattern: "^[0-9a-zA-Z]{1,8}$"
      minLength: 1
      maxLength: 8
      example: "0aB9wXyZ"
...
    Username:
      description: User name
      type: string
      pattern: "^[ -~]{1,64}$"
      minLength: 1
      maxLength: 64
      example: "Merchant-001"
    UUID:
      description: UUID
      type: string
      format: uuid
      pattern: "^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}$"
      minLength: 36
      maxLength: 36
      example: "d0d654e6-97da-4848-b568-99fedccb642b"

    # ------------------------------------------------------
    # Complex types
    # ------------------------------------------------------
    AccessToken:
      description: Data returned when a new access token is requested
      type: object
      additionalProperties: false
      properties:
        access_token:
          $ref: '#/components/schemas/Token'
        refresh_token:
          $ref: '#/components/schemas/Token'
        token_type:
          $ref: '#/components/schemas/BearerTokenType'
        expires_in:
          $ref: '#/components/schemas/PositiveInteger'
      required:
        - access_token
        - token_type
        - expires_in
      example:
        access_token: "eyJraWQiOiI2OWQzZDEwZi03MGRhLTQ5ODItYjI4ZC03MmU4MTQ4MzI4YTMiLCJhbGciOiJSUzI1NiJ9.eyJpc3MiOiJodHRwczovL21pbC1kLWFwaW0uYXp1cmUtYXBpLm5ldC9taWwtaWRwIiwiYXVkIjpbImh0dHBzOi8vbWlsLWQtYXBpbS5henVyZS1hcGkubmV0L21pbC1wYXltZW50LW5vdGljZSIsImh0dHBzOi8vbWlsLWQtYXBpbS5henVyZS1hcGkubmV0L21pbC1mZWUtY2FsY3VsYXRvciJdLCJleHAiOjE2Nzg3ODkyOTUsInNjb3BlIjoidmVyaWZ5QnlRckNvZGUgYWN0aXZhdGVCeVFyQ29kZSB2ZXJpZnlCeVRheENvZGVBbmROb3RpY2VOdW1iZXIgYWN0aXZhdGVCeVRheENvZGVBbmROb3RpY2VOdW1iZXIgY2xvc2UgZ2V0UGF5bWVudFN0YXR1cyBnZXRGZWUifQ.UJ8Vwv0NwF5S3fajgsdpgV-t6VeZAVC9dR0QGVpevC6Rw35tTx_1pJq2kI0GVAp0Vm7mNmgpZgzqx14QJfBmtTz3ZMJuUYVJUfHHQxxRxEyVBslrloOCqc8LfFkC-jmVQG5_9p9f6Tb7MB-h9J_f9ivtFX9gz7jK2Q_llD3HU0FujMgHEJiloKl7sp2uDaMIiy9dOKT7m0M6sA1tA84kwuJASudd_KWotideFsiA8LMBy_fGlEO7ogF-WytXD0XfW6VIBwQKq5KB3L3HxUuOZE1Gr2hubZ9SCeGAptiaPCWeEDf4w4x3IvKUL-rmm4bs4xalCmUbWcnTkKJi5f4N_8d2aRBrCL7k01rEK-OI1NH1-cCCk97eanNxvFEAKk_9t8rggipYt9BuL929p11LHVeYqqfemz1yrzc58LJXWPNgSMZ8XjKoITzsEiKWtGMYNV6GodJs_USnMKGIlc_uynZS0Q9mUV5IWhFaOZWFkM1nNwebMq9edg-kxtj1WpNkXfP1lahOMt36a9Rn2rWBLAhMBnerigsxFmUpi90EmbDkgOKTEXd3adOWLBb0jQSw0gPlNHe4bHh_LxBNNWljd6P9yZ4CYvZsWFwt_85d-PS56y-mNWV5oFMgPd-_qeF3Fn-QMj3Q5qySoqNCLzX7pnqjzXrveazLcR-CrroIUGM"
        refresh_token: "eyJraWQiOiIxN2YwNWViZC1hMWMzLTRjNWMtODM4Ny1jOWZlNjU4MDRiMDEiLCJhbGciOiJSUzI1NiJ9.eyJpc3MiOiJodHRwczovL21pbC1kLWFwaW0uYXp1cmUtYXBpLm5ldC9taWwtaWRwIiwiYXVkIjoiaHR0cHM6Ly9taWwtZC1hcGltLmF6dXJlLWFwaS5uZXQvbWlsLWlkcCIsImV4cCI6MTY3ODg4MTA1NSwic2NvcGUiOiJvZmZsaW5lX2FjY2VzcyJ9.Bu-BXKRMoE8fdBfuoPAa70WQYUWVH3BK3VFMw-9mr7hGlKBSONXQY8vcrFuh4OLm6vHX2hUagYuJ3yS_Anm5AHU8Kpay0dGm30fyPxFAyv5s_Yn_ET3VExPZBedBSlHl0Rzt0JQ_9X07X0JkCwWacxWMueKNoVndAh3QxswbkwtEANAKvQcvAtdLjh-93eDYFq0vvS8qNRReTZf46fldIL9H5-0MMqjfVi5A6y8YCzx7_kOHKq8wCy7u75QtKhPRzYnY0JcBVMALZgQovpCYpnlPkP3JV3cgdmlDGWmGHbKQz9iQ5bY1071e3vhxqR2iEqFJwn7nohYq9sxkMQGxNxz2BI-w3v_SQ3zS3mI1qK_qm1c7yzyF3c_lB920T5hBup9RozIXVb2OkX38pW6JKORETgSLlhnqLc8Rr1_HaQ2qN_JYcn7Urv_0mSJh-i8zS39F1glY3ejHHiVbl8cSJsPPw-ehpKAozf6h0_JiZBe0TuBtyTNULQH1iri19C9z_XBgBnfC1DtrnQn-UIgGbJq1TnIgtVmtTnReJqDpcf6_MGDLqCL-x192UQszb8GgWjD8tRusJ9X5HG1RHUAVyQ5gfOCegHmymyB7Dx7e5fCOghDUTT0t2dvVKbIXZGilrddKBg70iBGDA5j5-01lvnl2aaxBv51Y_zr-3IZdAyA"
        token_type: "Bearer"
        expires_in: 36000
...
    CreateClientReq:
      description: Request to create a new client
      additionalProperties: false
      properties:
        channel:
          $ref: '#/components/schemas/Channel'
        description:
          $ref: '#/components/schemas/ClientDescription'
        subject:
          $ref: '#/components/schemas/Subject'
      required:
        - description
      example:
        channel: "ATM"
        description: "ATM Layer - Test Environment"
    CreateClientRes:
      description: Response to a request to create a new client
      additionalProperties: false
      properties:
        clientId:
          $ref: '#/components/schemas/UUID'
        clientSecret:
          $ref: '#/components/schemas/ClientSecret'
      required:
        - clientId
        - clientSecret
      example:
        clientId: "d0d654e6-97da-4848-b568-99fedccb642b"
        clientSecret: "265a5ac0-c651-428e-bed6-763f6cd244b5"
...
    Errors:
      description: List of errors
      type: object
      additionalProperties: false
      properties:
        errors:
          type: array
          maxItems: 32
          items:
            $ref: '#/components/schemas/ErrorCode'
        descriptions:
          type: array
          maxItems: 32
          items:
            $ref: '#/components/schemas/ErrorDescription'
      required:
        - errors
      example:
        errors:
          - "00000000A"
          - "001000001"
        descriptions:
          - "[00000000A] Duplicate client id"
          - "[001000001] Invalid location"
    GetAccessTokenByClientCredentials:
      description: Request to get an access token by means of client credentials
      type: object
      additionalProperties: false
      properties:
        grant_type:
          $ref: '#/components/schemas/ClientCredentialsGrantType'
        client_id:
          $ref: '#/components/schemas/UUID'
        client_secret:
          $ref: '#/components/schemas/ClientSecret'
      required:
        - grant_type
        - client_id
        - client_secret
      example:
        grant_type: "client_credentials"
        client_id: "d0d654e6-97da-4848-b568-99fedccb642b"
        client_secret: "265a5ac0-c651-428e-bed6-763f6cd244b5"
    GetAccessTokenByExternalToken:
      description: Request to get an access token by means of external token
      type: object
      #additionalProperties: true
      properties:
        grant_type:
          $ref: '#/components/schemas/ExternalTokenGrantType'
        ext_token:
          $ref: '#/components/schemas/ExternalToken'
        add_data:
          $ref: '#/components/schemas/AdditionalData'
        scope:
          $ref: '#/components/schemas/OfflineAccessScope'
        client_id:
          $ref: '#/components/schemas/UUID'
      required:
        - grant_type
        - ext_token
        - add_data
        - client_id
      example:
        grant_type: "poynt_token"
        ext_token: "eyJhbGciOiJSUzI1NiJ9.eyJzdWIiOiJOZXhpIiwicG95bnQuZGlkIjoidXJuOnRpZDo1NTYyYjhlZC1lODljLTMzMmEtYThkYy1jYTA4MTcxMzUxMTAiLCJwb3ludC5kc3QiOiJEIiwicG95bnQub3JnIjoiMGU2Zjc4ODYtMDk1Ni00NDA1LWJjNDgtYzE5ODY4ZDdlZTIyIiwicG95bnQuc2N0IjoiVSIsImlzcyI6Imh0dHBzOlwvXC9zZXJ2aWNlcy1ldS5wb3ludC5uZXQiLCJwb3ludC51cmUiOiJPIiwicG95bnQua2lkIjozOTMyNDI1MjY4MDY5NDA5MjM0LCJwb3ludC5zY3YiOiJOZXhpIiwicG95bnQuc3RyIjoiZDNmZDNmZDMtMTg5ZC00N2M4LThjMzYtYjY4NWRkNjBkOTY0IiwiYXVkIjoidXJuOnRpZDo1NTYyYjhlZC1lODljLTMzMmEtYThkYy1jYTA4MTcxMzUxMTAiLCJwb3ludC51aWQiOjM3MzY1NzQsInBveW50LmJpeiI6IjRiN2ViOTRiLTEwYzktNGYxMS1hMTBlLTcyOTJiMjlhYjExNSIsImV4cCI6MTY4MDc4MzUzNiwiaWF0IjoxNjgwNjk3MTM2LCJqdGkiOiI3MjBmMDFlZS1iZDk4LTRjYjItOTU2Mi0xZjI4YWY2NGJhZGYifQ.dTuvtzqy9oPWIN4NIBdhIR09Xpm70dgRCP-ybuVLo24DwqaysPKNmEHDXcq2gGE1w2L6e783_PXRK3RI0j1TQRFeLRbiPzN5imBdrJ2LlV8QNdkElOl2x32j652YeFcoAitBzFss_Do0_rquU_008eeIXWa-B-AiMsdAqgLUiMigsTT42rQYr7Mb8Am_NWwvZ9-DWiox6HbuUNUo3TStBmLervqlQ5j2_3AzcOILp8cJX0699fw7Y6gcu_pNHgjswqD0UVPSAmHf_bqFAH6b98qHVKe3isMSoktYi4FfWdpG1ykviEp9Ii0QKfeLnqyIR6g2o2XzGlDv7Usv5ouiXg"
        add_data: "4b7eb94b-10c9-4f11-a10e-7292b29ab115"
        scope: "offline_access"
        client_id: "d0d654e6-97da-4848-b568-99fedccb642b"
    GetAccessTokenByPassword:
      description: Request to get an access token by means of resource owner password
      type: object
      #additionalProperties: true
      properties:
        grant_type:
          $ref: '#/components/schemas/PasswordGrantType'
        username:
          $ref: '#/components/schemas/Username'
        password:
          $ref: '#/components/schemas/Password'
        scope:
          $ref: '#/components/schemas/OfflineAccessScope'
        client_id:
          $ref: '#/components/schemas/UUID'
      required:
        - grant_type
        - username
        - password
        - client_id
      example:
        grant_type: "password"
        username: "Merchant-001"
        password: "Password-001$"
        scope: "offline_access"
        client_id: "d0d654e6-97da-4848-b568-99fedccb642b"
...
    RefreshAccessToken:
      description: Request to get an access token by means of a refresh token
      type: object
      #additionalProperties: true
      properties:
        grant_type:
          $ref: '#/components/schemas/RefreshGrantType'
        client_id:
          $ref: '#/components/schemas/UUID'
        refresh_token:
          $ref: '#/components/schemas/Token'
      required:
        - grant_type
        - client_id
        - refresh_token
      example:
        grant_type: "refresh_token"
        client_id: "d0d654e6-97da-4848-b568-99fedccb642b"
        refresh_token: "eyJraWQiOiIxN2YwNWViZC1hMWMzLTRjNWMtODM4Ny1jOWZlNjU4MDRiMDEiLCJhbGciOiJSUzI1NiJ9.eyJpc3MiOiJodHRwczovL21pbC1kLWFwaW0uYXp1cmUtYXBpLm5ldC9taWwtaWRwIiwiYXVkIjoiaHR0cHM6Ly9taWwtZC1hcGltLmF6dXJlLWFwaS5uZXQvbWlsLWlkcCIsImV4cCI6MTY3ODg4MTA1NSwic2NvcGUiOiJvZmZsaW5lX2FjY2VzcyJ9.Bu-BXKRMoE8fdBfuoPAa70WQYUWVH3BK3VFMw-9mr7hGlKBSONXQY8vcrFuh4OLm6vHX2hUagYuJ3yS_Anm5AHU8Kpay0dGm30fyPxFAyv5s_Yn_ET3VExPZBedBSlHl0Rzt0JQ_9X07X0JkCwWacxWMueKNoVndAh3QxswbkwtEANAKvQcvAtdLjh-93eDYFq0vvS8qNRReTZf46fldIL9H5-0MMqjfVi5A6y8YCzx7_kOHKq8wCy7u75QtKhPRzYnY0JcBVMALZgQovpCYpnlPkP3JV3cgdmlDGWmGHbKQz9iQ5bY1071e3vhxqR2iEqFJwn7nohYq9sxkMQGxNxz2BI-w3v_SQ3zS3mI1qK_qm1c7yzyF3c_lB920T5hBup9RozIXVb2OkX38pW6JKORETgSLlhnqLc8Rr1_HaQ2qN_JYcn7Urv_0mSJh-i8zS39F1glY3ejHHiVbl8cSJsPPw-ehpKAozf6h0_JiZBe0TuBtyTNULQH1iri19C9z_XBgBnfC1DtrnQn-UIgGbJq1TnIgtVmtTnReJqDpcf6_MGDLqCL-x192UQszb8GgWjD8tRusJ9X5HG1RHUAVyQ5gfOCegHmymyB7Dx7e5fCOghDUTT0t2dvVKbIXZGilrddKBg70iBGDA5j5-01lvnl2aaxBv51Y_zr-3IZdAyA"
...
  requestBodies:
    CreateClient:
      description: Request to create e new client
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateClientReq'
    GetAccessToken:
      description: Request to get an access token by means of username and password or by refresh token or by external token
      content:
        application/x-www-form-urlencoded:
          schema:
            oneOf:
              - $ref: '#/components/schemas/GetAccessTokenByPassword'
              - $ref: '#/components/schemas/RefreshAccessToken'
              - $ref: '#/components/schemas/GetAccessTokenByExternalToken'
              - $ref: '#/components/schemas/GetAccessTokenByClientCredentials'
...
  parameters:
    AcquirerId:
      name: AcquirerId
      in: header
      description: Acquirer ID assigned by PagoPA
      required: false
      schema:
        $ref: '#/components/schemas/AcquirerId'
    Channel:
      name: Channel
      in: header
      description: Channel originating the request
      required: false
      schema:
        $ref: '#/components/schemas/Channel'
    FiscalCode:
      name: FiscalCode
      in: header
      description: User fiscal (tax) code
      required: false
      schema:
        $ref: '#/components/schemas/FiscalCode'
    MerchantId:
      name: MerchantId
      description: Merchant ID. Mandatory when Channel equals POS.
      in: header
      required: false
      schema:
        $ref: '#/components/schemas/MerchantId'
    PageNumber:
      name: page
      in: query
      description: Number of the requested page of data
      required: true
      schema:
        $ref: '#/components/schemas/PageNumber'
    PageSize:
      name: size
      in: query
      description: Size of the requested page of data
      required: true
      schema:
        $ref: '#/components/schemas/PageSize'
    RequestId:
      name: RequestId
      in: header
      description: Request Id that will be logged by services
      required: false
      schema:
        $ref: '#/components/schemas/UUID'
    TerminalId:
      name: TerminalId
      in: header
      description: ID of the terminal originating the transaction. It must be unique per acquirer and channel.
      required: false
      schema:
        $ref: '#/components/schemas/TerminalId'
    Version:
      name: Version
      in: header
      description: Version of the required API
      required: false
      schema:
        type: string
        pattern: "^[ -~]{1,64}$"
        minLength: 1
        maxLength: 64
        example: "1.0.0-alpha-a.b-c-somethinglong+build.1-aef.1-its-okay"

  # ========================================================
  # Responses
  # ========================================================
  responses:
    AccessToken:
      description: Response returned when an access token is requested
      headers:
        Access-Control-Allow-Origin:
          description: Indicates whether the response can be shared with requesting code from the given origin
          required: false
          schema:
            $ref: '#/components/schemas/AccessControlAllowOrigin'
        RateLimit-Limit:
          description: The number of allowed requests in the current period
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitLimit'
        RateLimit-Reset:
          description: The number of seconds left in the current period
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitReset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AccessToken'
    CreateClient:
      description: Response returned when the creation of a new client is requested
      headers:
        Access-Control-Allow-Origin:
          description: Indicates whether the response can be shared with requesting code from the given origin
          required: false
          schema:
            $ref: '#/components/schemas/AccessControlAllowOrigin'
        RateLimit-Limit:
          description: The number of allowed requests in the current period
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitLimit'
        RateLimit-Reset:
          description: The number of seconds left in the current period
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitReset'
        Location:
          description: URL of the client
          required: true
          schema:
            $ref: '#/components/schemas/ClientLocation'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateClientRes'
...
    Error:
      description: Error response
      headers:
        Access-Control-Allow-Origin:
          description: Indicates whether the response can be shared with requesting code from the given origin
          required: false
          schema:
            $ref: '#/components/schemas/AccessControlAllowOrigin'
        RateLimit-Limit:
          description: The number of allowed requests in the current period
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitLimit'
        RateLimit-Reset:
          description: The number of seconds left in the current period
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitReset'
        Retry-After:
          description: The number of seconds to wait before allowing a follow-up request
          required: false
          schema:
            $ref: '#/components/schemas/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
        text/*:
          schema:
            type: string
            pattern: "^[ -~]{0,65535}$"
            maxLength: 65535
...
    PageOfClients:
      description: Response to the request to get clients
      headers:
        Access-Control-Allow-Origin:
          description: Indicates whether the response can be shared with requesting code from the given origin
          required: false
          schema:
            $ref: '#/components/schemas/AccessControlAllowOrigin'
        RateLimit-Limit:
          description: The number of allowed requests in the current period
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitLimit'
        RateLimit-Reset:
          description: The number of seconds left in the current period
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitReset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PageOfClients'
...
  securitySchemes:
    oAuth2:
      description: A bearer token in the format of a JWS and conforms to the specifications included in RFC8725
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: /token
          refreshUrl: /token
          scopes:
            token_info: introspect access token
            admin: admin operations